Seasonal Crop Implementation
Crops' growing speed now affected by seasonal variation. Certain crops grow faster in certain seasons. Any crops which are not suited to winter have a 1 in 20 chance of dying each night due to extreme cold. Ground work is in place for crops to have a visible indicator when they have a chance to die.